requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
htmldocx==0.0.6
idna==3.4
lxml==4.9.3
markdown2==2.4.10
multidict==6.0.4
numpy==1.25.2
openai==0.28.0
pandas==2.1.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tenacity==8.2.3
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
